Changed order of loops: flights is now the outer loop and individuals is nested inside.  This allows all individuals to train on same sequence of flights.  Also, reduces number of new real populations that are created -- now one per flight rather than once per individual per flight.  Minor fix in real_players -- cast dude_num to int.

+++ real_players.py
@@ -22,7 +22,7 @@ def getRealMix(length):
     l.extend(['business' for i in range(business_num)])
 
     # there are likely to be a few passengers who will accept any offer
-    dude_num = math.floor(length * .01)
+    dude_num = int(math.floor(length * .01))
     l.extend(['dude' for i in range(dude_num)])
 
     # num_left = length - len(l)
